backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.1.0
python-multipart==0.0.6
# Starlette TestClient uses `httpx.Client(app=...)`; httpx 0.28 removed the `app` kwarg.
httpx>=0.26.0,<0.28.0

# OpenAI
openai>=1.40.0

# Anthropic (for Kimi.com and Claude)
anthropic>=0.75.0

# LangGraph for agentic orchestration
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Static Analysis
bandit==1.7.5

# Dynamic Testing (Optional)
hypothesis==6.92.2

# Cryptography
cryptography==41.0.7

# Database
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pyyaml>=6.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.7
